[feat] refactor, #8
